Add sub architecture to p_to_r_init
the hardware probe p_to_r_init is intentionally
different from the software probe one because of
the hardware probe listing information on /dev/sda.
/dev/sda is a USB stick for hardware probes, but
for software probes may be the system harddrive,
which we do not want to report about. It would
not hurt if we would, but could be privacy
sensitive.