feat: saved RI network information (static), for boot/upgrade
When the client initially connects, it will use DHCP as there is no
saved network configuration. This is problematic as network configuration
is not preserved during system upgrades or reboots, and this can cause
issues if the probe requires static configuration.

Note: this does not allow for fallback on DHCP if static is broken.