feat: add netconfig DNS  parsing to hwprobe
DNS static configuration is different from other static configuration.
Unlike IPv4 or IPv6, the format is meant for resolv.conf, but we want
something that is reginit-style.

```
nameserver 127.0.0.1
nameserver ::1
```

As such, it requires different parsing and reworking,
We remove all newlines and 'nameserver's, we add 'DNS_SERVERS' at the
start, and add a newline at the end. This brings it in line with the
expected reginit format.