add: atlasinit.c IPv6 envvars to file
For IPv4, 'atlasinit.c' prints environment variables to file, this
functionality is missing for IPv6. This commit fixes that.

This enables the parsing of all netconfig files for saved network settings:
- IPv4
- IPv6
- DNS (both IPv4/IPv6)

For context, atlasinit.c parses the reginit responses.
For IPv4 and IPv6, it parses the response into netconfig files, which contain
both **commands** and environment variables.
For DNS, it parses the response into a resolv.conf file, which is relatively
easy to parse, even if it does not contain environment variables.

Instead of rewriting large parts of complex code, we rely upon this "saving
to file" feature, and will use the environment variables and easy-to-parse
files to generate reginit-style networking configuration on-the-fly for
the openwrt-hwprobe arch.