feat: use netconfig as a static source for hwprobe
The alternative to refactoring the atlas loop / reginit is to generate
reginit-style lines (imitating the regserver), based on the netconfig files
generated by 'atlasinit.c'.

From my understanding, the reginit script will save and parse the networking
indications from the regserver through atlasinit.c, which is called by reginit,
as such, that is the source of truth.

As the prev. commit "rm:   saved RI network information (static)" mentions:

> Generating a new custom file is too time-intensive, as the main scripts
> (main loop/reginit) call 'atlasinit.c' to do network initialization, using
> netconfig files.
>
> Reworking the main loop and reginit, while possible and necessary, would
> currently require too much time and would be too invasive.

---

Why this focus on reginit-style lines?

We try to minimize the number of changes, KISS.

Michel Stam, the previous maintainer, had started work on a new architecture
('/bin/arch/') 'openwrt-hwprobe'. This was supposed to be a unifying
architecture for all RIPE Atlas hardware probes, in the goal of simplifying
the software probe code, by abstracting away hardware-specifics through
so-called '_wrt_syscall's, which are actually ubus calls (OpenWrt's micro bus).

Those ubus calls would then translate to invokations of hwprobe or
hwprobe-version-specfic scripts, which would perform those tasks.
Those scripts are registered via procd as hotplug scripts, thus the call to
the 'hotplug.ripe-atlas' object in ubus.

These are changes to the 'openwrt-hwprobe' architecture, which is not
feature-complete, but the existing code has been tested and is confirmed to
work.

The existing code includes parsing of network configuration from a reginit
reply, and applying it to OpenWrt, via those ubus calls. This is done via the
following shell functions:
- '_ri_parse'
- '_ri_value'
- '_handle_dns'
- 'set_ipv4'
- 'set_ipv6'