fix: invalid NETCONFIG_V*_DEST before reginit
The values of NETCONFIG_V*_DEST change over time.
More specifically, they move from /etc/, to ATLAS_SYSCONFDIR, /etc/ripe-atlas/.

As a result, during bootup, the file checking will be incorrect, which
causes the static configuration not to be loaded, and DHCP to be used.

This is not a great fix, as it duplicates the definition of netconfig paths,
but it is the approach chosen by generic-reginit.sh:21-22, instead of fixing
it at the source: common.sh.in:13-15.

Ideally, this would be fixed at the source, but that would require core changes
I am not fully comfortable making within a simple MR to add static networking.

---

Initial boot values are incorrect (NETCONFIG_V*_DEST only)

> Tue Nov 25 00:43:56 2025 user.debug ripe-atlas: RUN-TIME TIME: NETCONFIG_V4_DEST="/etc/netconfig_v4.sh", NETCONFIG_V6_DEST="/etc/netconfig_v6.sh", RESOLV_CONF_STATIC="/etc/ripe-atlas/resolv.conf.static"
> ...
> Tue Nov 25 00:43:56 2025 daemon.info ripe-atlas[2369]: start reg
> Tue Nov 25 00:43:56 2025 user.debug ripe-atlas: IMPORT TIME: NETCONFIG_V4_DEST="/etc/ripe-atlas/netconfig_v4.sh", NETCONFIG_V6_DEST="/etc/ripe-atlas/netconfig_v6.sh", RESOLV_CONF_STATIC="/etc/ripe-atlas/resolv.conf.static"
> Tue Nov 25 00:43:56 2025 daemon.info ripe-atlas[2369]: ATLAS registration starting

These NETCONFIG_V*_DEST variables are typically set within (arch)-reginit.sh,
in our case the correct value is stored in generic-reginit.sh, but the values
differ per-architecture. The core issue is that reginit.sh is only called later
(l.248 in ripe-atlas), after setting v4 and v6. see generic-reginit.sh:21-22.

Instead, the problem is caused by running SETUP_NETWORK_CMD (which calls
set_ipv4 and set_ipv6) BEFORE reginit is called. (rioe-atlas:81)

---

Background

The vars are initially set in common.sh.in:13-15 to /etc/, and then modified to
/etc/ripe-atlas/ in generic-reginit.sh.in:21-22. But as reginit is only called
much later in ripe-atlas.in:248, after the first call to SETUP_NETWORK_CMD,
they will initially have incorrect /etc/ values (set by common.sh.in:13-15),
instead of the correct /etc/ripe-atlas ones.