MR: add (saved) static network configuration to openwrt-hwprobe
This MR adds parsing of saved static network configuration to openwrt-hwprobe,
allowing a probe to boot, even without a reginit_reply. This MR also includes
additions to `atlasinit.c` to add IPv4 environment variable generation.

---

The choice of using the pre-existing `atlasinit.c`-generated netconfig files
is conscious, as the alternative is refactoring the atlas loop / reginit.
This would be too invasive and time-intensive.

One might think "just save and re-use the values in this script?", but
`set_ipv4` and `set_ipv6` are not called after each reginit reply, unlike
`atlasinit.c` which parses values within reginit replies.

From my understanding, the reginit script will save and parse (among other
things) the networking indications from the regserver through atlasinit.c,
which is called by reginit, as such, those netconfig/resolv.conf files are
the source of truth.

---

Why this focus on reginit-style lines?

We try to minimize the number of changes, KISS.

Michel Stam, the previous maintainer, had started work on a new architecture
(`/bin/arch/`) `openwrt-hwprobe`. This was supposed to be a unifying
architecture for all RIPE Atlas hardware probes, in the goal of simplifying
the software probe code, by abstracting away hardware-specifics through
so-called `_wrt_syscall`s, which are actually ubus calls (OpenWrt's micro bus).

Those ubus calls would then translate to invocations of hwprobe or
hwprobe-version-specfic scripts, which would perform those tasks.
Those scripts are registered via procd as hotplug scripts, thus the call to
the `hotplug.ripe-atlas` object in ubus.

These are changes to the `openwrt-hwprobe` architecture, which is not
feature-complete, but the existing code has been tested and is confirmed to
work.

The existing code includes parsing of network configuration from a reginit
reply, and applying it to OpenWrt, via those ubus calls. This is done via the
following shell functions:
- `_ri_parse`
- `_ri_value`
- `_handle_dns`
- `set_ipv4`
- `set_ipv6`

Link: GL-152
Reviewed-by: Marios Mavropoulos <mmavropoulos@ripe.net>